refactor: follow api 함수 코드 한 파일로 통합
